refactor: 🚚 classify and organize classes
