+++ src/main/java/com/zhanganzhi/chathub/ChatHub.java
@@ -7,9 +7,9 @@ import com.velocitypowered.api.event.proxy.ProxyShutdownEvent;
 import com.velocitypowered.api.plugin.Plugin;
 import com.velocitypowered.api.plugin.annotation.DataDirectory;
 import com.velocitypowered.api.proxy.ProxyServer;
-import com.zhanganzhi.chathub.command.Command;
 import com.zhanganzhi.chathub.core.Config;
 import com.zhanganzhi.chathub.core.EventHub;
+import com.zhanganzhi.chathub.platforms.velocity.VelocityCommand;
 import org.slf4j.Logger;
 
 import java.nio.file.Path;
@@ -56,7 +56,7 @@ public class ChatHub {
         // command
         proxyServer.getCommandManager().register(
                 proxyServer.getCommandManager().metaBuilder("chathub").plugin(this).build(),
-                new Command(this)
+                new VelocityCommand(this)
         );
 
         // init event hub
