+++ src/main/java/com/zhanganzhi/chathub/platforms/Platform.java
@@ -1,4 +1,4 @@
-package com.zhanganzhi.chathub.entity;
+package com.zhanganzhi.chathub.platforms;
 
 public enum Platform {
     DISCORD("discord"),
