+++ src/main/java/com/zhanganzhi/chathub/platforms/kook/KookDaemon.java
@@ -1,4 +1,4 @@
-package com.zhanganzhi.chathub.adaptors.kook;
+package com.zhanganzhi.chathub.platforms.kook;
 
 import com.zhanganzhi.chathub.core.Config;
 import org.slf4j.Logger;
